Refactor: Avoid circular imports for framework version
